+++ chrome_logger/__init__.py
@@ -1,3 +1,3 @@
 """Chrome Network Logger package."""
 
-__version__ = "3.1.0"
+__version__ = "3.1.1"